Allow user to edit newsletter entries
On login to admin page, users can now submit information for new
newsletters.

Modified the controllers for the admin and newsletters pages, added newsletters as a
resource. Added a form for new newsletters, included it on the admin page.
Added a helper method to get newsletter names.

Still to be done: add file uploads to newsletter additions, specify which
users are able to edit newsletters, create a nicer view to display them.