Add database config file for dev to ease setup
The reference to the .openshift directory doesn't work for running
locally. With this in mind, an extra db config file was originally
created (database_prod.yml) with the idea in mind that the two would
switch off. It happens that the main file was carelessly overwritten
before so now both have prod style db configs.

Putting the dev file here also makes it easy to read into the main file as
needed, without the need to recommit a main file that will cause
deployments to fail.