Add styling to audio player for Sermons
Increase width to take up most of the section, so it doesn't seem like as much wasted space. Create a light colored box shadow, so that the player shows up against the dark colored background. Outline for player removed as it's visually distracting. However, some signifier should be added later to tell when player is in focus.