Add test for home page view
Added this first view test to make sure that actually rendering the view doesn't hit an error. It could be easy to miss a syntax error in a view somewhere. This test should also catch situations like invalid db calls. Importantly, it also renders the view within the main application layout, so problems with the layout will get caught too.