First pass at adding permission validation
This is causing more tests to fail, but it's one step closer to better
validations. Previously, newsletters controller only checked to make sure
a user was logged in, did not look at their permissions. That's not so
bad, as the admin page will still hide views as needed. However, it would
be better if create, destroy, edit actions weren't available without
correct permissions.