Make Newsletters index more dynamic
Added expand/collapse to the Newsletter headings for each year. Also added anchor links to the sidebar that both go to the year, and expand that section. Minor problem here though: the position the anchor goes to is fixed before the section expands, which changes how far down the page the anchor is able to go. This is minor enough that I'm ignoring for now.

Reorganized some styles so that I have a consistent class for sidebar navs and one for interactable elements. Made sidebar links on the History page match the ones on the Newsletter page, and all of those also match the Newsletter year headings.