Make dropdowns look better
Had some issues figuring out exactly how to specify classes on some of the
dropdowns. That's been worked out, so the right classes are assigned.
Added a class for date selectors, so that they'll take ~1/3 of the width
and line up horizontally. Added form-group containers for the other select
boxes, so their bottom margins work out correctly.