Align logout button, reorganize css for buttons
Logout button on admin page was not lined up evenly with the Admin heading to its left. Set flexbox properties to get these to look right.

The class for buttons in the style currently used in sidebars was tied up with list styling. Moved some properties out into a button-cell class.