Add a single column layout for 320-479 width views
Included media queries to show home page lower sections in a single column
for smaller devices. Hoping this fixes the readibility for smaller phones.

On the downside, I'm hiding/showing just with display:none, so hidden
sections are still loaded. The content of some sections, especially the
placeholder area for events, is repeated. This info needs to be
consolidated into partials to clean up the home page's view file.