Make sidebar nav buttons less overwhelming
On the History and Newsletter pages, there are buttons to jump down to
different years. On smaller viewports, the sidebar would collapse into the
one column and show a long line of very wide buttons.

For small devices, the buttons will now go down to a smaller width and fit
3 per row.

Also lowered the breakpoint for the newsletter/history pages' column
layouts. Was showing 2 column layout for medium and large, now 2 columns
are for small (according to Bootstrap) and larger.

Added a bit to handle hyphenation for the Newsletter page's header. That
heading was too long for the container.