Add Event.upcoming_and_recent
This method will be used to get data for the homepage.

Also modified the test for this, as it became apparent that we probably
don't want to remove events as soon as they happen. Might need to
eventually prune out recent ones, on the condition that future events are
available?