Show start and end dates for events in views
Summary on home page now includes end dates for events that have them. The
views for creating and editing events will show these correctly as well.
So at this point there's a little flexibility in dates for events.

Also added methods to the Event model to get formatted start and end dates
(MMM DD), as well as tests for those methods.