Allow events to expand to show details
Descriptions for events were previously not accessible. These have now been
added in the summary table found on the homepage. Using plus and minus
glyphicons to indicate the available actions.

Also of note, changed the glyphicon import in the application.css file to
use https; this should prevent mixed content warnings for the homepage.