Handle expanding events in js file, not bootstrap
There were a couple of bugs in the previous update, that introduced
expanding and collapsing for event details.

First, clicking the expand button an even number of times quickly would
cause the class/icon to go out of sync. So double clicking would lead to a
plus icon with the info already expanded. To resolve this, I included a
jQuery slideToggle in the same event handling that changes the icon (and
removed the Bootstrap collapse classes). This way, the icon should stay in sync.

The other issue encountered was javascript not reloading after loading new
pages. Toggling plus/minus icons would work correctly at initial page
visit or refresh, but after navigating to a different page and returning,
would not work. This is an issue with Turbolinks, and the solution is to
listen for page change events, then restore the event listener for the
buttons. This also fixes a consistency error with the collapsable sections
of the Newsletters page.