Add utility to change video displayed on home page
We're starting to upload videos, and at least for now a prominent piece of the home page features a video of the latest service. For now, just need a quick and simple way to update that video without a code change. We only need the id to change what's shown, so that's all the field expects. For simplicity's sake, this is just saved in a local file. this gives some flexibility, in that the video doesn't need to be from a sermon record, or necessarily from our channel even.

In the future this will probably need to do more filtering, right now it falls back to a default video if the input is bad. But ideally it should filter a full video view or embed URL into just a video ID.